Raw text in dataset, tokenizing in sender/receiver
In order to avoid translation from one word embedding to the other the sentences are fed raw to the sender and receiver. It becomes their duty to tokenize the words. 
As per now sentence splitting (spaces, commas, numbers) is carried out independently by the two, a common preprocessing step could also be implemented.